+++ src/EmbDB/EmbDB/BPLeafNodeMapROSimpleCompressor.h
@@ -53,7 +53,7 @@ namespace embDB
 			return true;
 		}
 	private:
-		size_t m_nSize;
+		uint32 m_nSize;
 	};
 }
